backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5
pydantic==2.6.1
pydantic-settings==2.1.0
openai>=1.30.0
httpx==0.26.0
beautifulsoup4==4.12.3
pdfminer.six==20221105
tabula-py==2.9.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
lxml==5.1.0
chardet==5.2.0
numpy==1.26.4
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
greenlet==3.0.3
